Implemtacion de metodos para recuperar la orden en estado c al iniciar secion

+++ src/main/java/com/fsociety/storeservices/service/impl/TshopoingcartServiceImpl.java
@@ -1,12 +1,14 @@
 package com.fsociety.storeservices.service.impl;
 
 import com.fsociety.storeservices.entity.Torders;
+import com.fsociety.storeservices.entity.Tproducts;
 import com.fsociety.storeservices.entity.Tshopoingcart;
 import com.fsociety.storeservices.entity.bo.OrderFullBuilder;
 import com.fsociety.storeservices.entity.bo.OrdersFullBO;
 import com.fsociety.storeservices.entity.bo.ShoppingBuilder;
 import com.fsociety.storeservices.entity.bo.ShoppingCartBO;
 import com.fsociety.storeservices.repository.TordersRepository;
+import com.fsociety.storeservices.repository.TproductsRepository;
 import com.fsociety.storeservices.repository.TshopoingcartRepository;
 import com.fsociety.storeservices.service.TshopoingcartService;
 import org.slf4j.Logger;
@@ -35,11 +37,30 @@ public class TshopoingcartServiceImpl implements TshopoingcartService {
     @Autowired
     private TordersRepository tordersRepository;
 
+    @Autowired
+    private TproductsRepository tproductsRepository;
+
     @Override
     public void insert(Tshopoingcart tshopoingcart) throws Exception {
         LOGGER.debug(">>>Insert()->tshopoingcart:{}", tshopoingcart);
+        Tshopoingcart cart = null;
+        Tproducts tproducts = null;
+        Integer amountAdd = 0;
         try {
+            amountAdd = tshopoingcart.getAmount();
+            cart = tshopoingcartRepository.findShopoingcartByOrderAndProduct(tshopoingcart.getIdOrder().getId(), tshopoingcart.getIdProduct().getId());
+            if (cart != null) {
+                Integer amountActual = cart.getAmount();
+                tshopoingcart = cart;
+                tshopoingcart.setAmount(amountActual + amountAdd);
+            }
+            tproducts = tproductsRepository.getOne(tshopoingcart.getIdProduct().getId());
+            if (tproducts == null) {
+                throw new Exception("No se encuentra el porducto");
+            }
+            tproducts.setInventary(tproducts.getInventary() - amountAdd);
             tshopoingcart.setStatus(true);
+            tproductsRepository.save(tproducts);
             tshopoingcartRepository.save(tshopoingcart);
         } catch (Exception e) {
             LOGGER.error("Exception: {}", e);
@@ -56,16 +77,6 @@ public class TshopoingcartServiceImpl implements TshopoingcartService {
             if (!tshopoingcartOptional.isPresent()) {
                 throw new Exception("No existe el registro");
             }
-//			//idOrder
-//			if(data.containsKey("idOrder")){
-//				Integer idOrder = (Integer)data.get("idOrder");
-//				tshopoingcartOptional.get().setIdOrder(idOrder);
-//			}
-//			//idProduct
-//			if(data.containsKey("idProduct")){
-//				Integer idProduct = (Integer)data.get("idProduct");
-//				tshopoingcartOptional.get().setIdProduct(idProduct);
-//			}
             //amount
             if (data.containsKey("amount")) {
                 Integer amount = (Integer) data.get("amount");
@@ -86,14 +97,21 @@ public class TshopoingcartServiceImpl implements TshopoingcartService {
     @Override
     public void delete(Integer id) throws Exception {
         LOGGER.debug(">>>> delete->id: {}", id);
+        Tproducts tproducts = null;
         try {
             Optional<Tshopoingcart> tshopoingcartOptional = tshopoingcartRepository.findById(id);
             if (!tshopoingcartOptional.isPresent()) {
                 throw new Exception("No existe el registro");
             }
+            tproducts = tproductsRepository.getOne(tshopoingcartOptional.get().getIdProduct().getId());
+            if (tproducts == null) {
+                throw new Exception("Error al encontrar el producto");
+            }
+            tproducts.setInventary(tshopoingcartOptional.get().getAmount() + tproducts.getInventary());
             tshopoingcartRepository.delete(tshopoingcartOptional.get());
+            tproductsRepository.save(tproducts);
         } catch (Exception e) {
-            LOGGER.error("Exception: {}", e);
+            LOGGER.error("Exception: {}", e.getMessage());
             throw new Exception(e);
         }
     }
@@ -121,9 +139,9 @@ public class TshopoingcartServiceImpl implements TshopoingcartService {
 
         try {
             torders = tordersRepository.findOrderActiveByUserStateC(idOrder, idUser);
-            if(torders==null){
-            	throw new Exception("No existe la order");
-			}
+            if (torders == null) {
+                throw new Exception("No existe la order");
+            }
             shoppingCartBOS = new ArrayList<>();
             List<Tshopoingcart> tshopoingcarts = tshopoingcartRepository.findShoppingCartByOrder(torders.getId());
             AtomicReference<Double> sum = new AtomicReference<>(0.0);
